requirements.txt
paho_mqtt==1.6.1
setuptools==45.2.0
